move all the test builders to a separate class
