+++ src/native/managed/cdacreader/tests/RangeSectionMapTests.cs
@@ -6,15 +6,11 @@ using Xunit;
 
 using Microsoft.Diagnostics.DataContractReader.ExecutionManagerHelpers;
 
-using InteriorMapValue = Microsoft.Diagnostics.DataContractReader.ExecutionManagerHelpers.RangeSectionMap.InteriorMapValue;
 
 namespace Microsoft.Diagnostics.DataContractReader.UnitTests;
 
 public class RangeSectionMapTests
 {
-    const ulong DefaultTopLevelAddress = 0x0000_1000u; // arbitrary
-    const int EntriesPerMapLevel = 256; // for now its fixed at 256, see codeman.h RangeSectionMap::entriesPerMapLevel
-    const int BitsPerLevel = 8;
     internal class RSLATestTarget : TestPlaceholderTarget
     {
         private readonly MockTarget.Architecture _arch;
@@ -28,141 +24,13 @@ public class RangeSectionMapTests
         }
     }
 
-    internal class Builder
-    {
-        private readonly TargetPointer _topLevelAddress;
-        private readonly MockMemorySpace.Builder _builder;
-        private readonly TargetTestHelpers _targetTestHelpers;
-        private readonly int _levels;
-        private readonly int _maxSetBit;
-        private ulong _nextMapAddress;
-        public Builder(MockTarget.Architecture arch) : this (DefaultTopLevelAddress, new MockMemorySpace.Builder (new TargetTestHelpers(arch)))
-        {
-        }
-
-        public Builder (TargetPointer topLevelAddress, MockMemorySpace.Builder builder)
-        {
-            _topLevelAddress = topLevelAddress;
-            _builder = builder;
-            _targetTestHelpers = builder.TargetTestHelpers;
-            var arch = _targetTestHelpers.Arch;
-            _levels = arch.Is64Bit ? 5 : 2;
-            _maxSetBit = arch.Is64Bit ? 56 : 31; // 0 indexed
-            MockMemorySpace.HeapFragment top = new MockMemorySpace.HeapFragment
-            {
-                Address = topLevelAddress,
-                Data = new byte[EntriesPerMapLevel * _targetTestHelpers.PointerSize],
-                Name = $"Map Level {_levels}"
-            };
-            _nextMapAddress = topLevelAddress + (ulong)top.Data.Length;
-            _builder.AddHeapFragment(top);
-        }
-
-        public TargetPointer TopLevel => _topLevelAddress;
-
-        private int EffectiveBitsForLevel(ulong address, int level)
-        {
-            ulong addressBitsUsedInMap = address >> (_maxSetBit + 1 - (_levels * BitsPerLevel));
-            ulong addressBitsShifted = addressBitsUsedInMap >> ((level - 1) * BitsPerLevel);
-            int addressBitsUsedInLevel = checked((int)((EntriesPerMapLevel - 1) & addressBitsShifted));
-            return addressBitsUsedInLevel;
-        }
-
-        // This is how much of the address space is covered by each entry in the last level of the map
-        private int BytesAtLastLevel => checked (1 << BitsAtLastLevel);
-        private int BitsAtLastLevel => _maxSetBit - (BitsPerLevel * _levels) + 1;
-
-        private TargetPointer CursorAddress(RangeSectionMap.Cursor cursor)
-        {
-            return cursor.LevelMap + (ulong)(cursor.Index * _targetTestHelpers.PointerSize);
-        }
-
-        private void WritePointer(RangeSectionMap.Cursor cursor, InteriorMapValue value)
-        {
-            TargetPointer address = CursorAddress(cursor);
-            Span<byte> dest = _builder.BorrowAddressRange(address, _targetTestHelpers.PointerSize);
-            _targetTestHelpers.WritePointer(dest, value.RawValue);
-        }
-
-        private InteriorMapValue LoadCursorValue (RangeSectionMap.Cursor cursor)
-        {
-            TargetPointer address = CursorAddress(cursor);
-            ReadOnlySpan<byte> src = _builder.BorrowAddressRange(address, _targetTestHelpers.PointerSize);
-            return new InteriorMapValue(_targetTestHelpers.ReadPointer(src));
-        }
-
-        private MockMemorySpace.HeapFragment AllocateMapLevel(int level)
-        {
-            MockMemorySpace.HeapFragment mapLevel = new MockMemorySpace.HeapFragment
-            {
-                Address = new TargetPointer(_nextMapAddress),
-                Data = new byte[EntriesPerMapLevel * _targetTestHelpers.PointerSize],
-                Name = $"Map Level {level}"
-            };
-            _nextMapAddress += (ulong)mapLevel.Data.Length;
-            _builder.AddHeapFragment(mapLevel);
-            return mapLevel;
-        }
-
-
-        // computes the cursor for the next level down from the given cursor
-        // if the slot for the next level does not exist, it is created
-        private RangeSectionMap.Cursor GetOrAddLevelSlot(TargetCodePointer address, RangeSectionMap.Cursor cursor, bool collectible = false)
-        {
-            int nextLevel = cursor.Level - 1;
-            int nextIndex = EffectiveBitsForLevel(address, nextLevel);
-            InteriorMapValue nextLevelMap = LoadCursorValue(cursor);
-            if (nextLevelMap.IsNull)
-            {
-                nextLevelMap = new (AllocateMapLevel(nextLevel).Address);
-                if (collectible)
-                {
-                    nextLevelMap = new (nextLevelMap.RawValue | 1);
-                }
-                WritePointer(cursor, nextLevelMap);
-            }
-            return new RangeSectionMap.Cursor(nextLevelMap.Address, nextLevel, nextIndex);
-        }
-
-        // ensures that the maps for all the levels for the given address are allocated.
-        // returns the address of the slot in the last level that corresponds to the given address
-        RangeSectionMap.Cursor EnsureLevelsForAddress(TargetCodePointer address, bool collectible = false)
-        {
-            int topIndex = EffectiveBitsForLevel(address, _levels);
-            RangeSectionMap.Cursor cursor = new RangeSectionMap.Cursor(TopLevel, _levels, topIndex);
-            while (!cursor.IsLeaf)
-            {
-                cursor = GetOrAddLevelSlot(address, cursor, collectible);
-            }
-            return cursor;
-        }
-        public void InsertAddressRange(TargetCodePointer start, uint length, ulong value, bool collectible = false)
-        {
-            TargetCodePointer cur = start;
-            ulong end = start.Value + length;
-            do {
-                RangeSectionMap.Cursor lastCursor = EnsureLevelsForAddress(cur, collectible);
-                WritePointer(lastCursor, new InteriorMapValue(value));
-                cur = new TargetCodePointer(cur.Value + (ulong)BytesAtLastLevel); // FIXME: round ?
-            } while (cur.Value < end);
-        }
-        public void MarkCreated()
-        {
-            _builder.MarkCreated();
-        }
-
-        public MockMemorySpace.ReadContext GetReadContext()
-        {
-            return _builder.GetReadContext();
-        }
-    }
 
 
     [Theory]
     [ClassData(typeof(MockTarget.StdArch))]
     public void TestLookupFail(MockTarget.Architecture arch)
     {
-        var builder = new Builder(arch);
+        var builder = new ExecutionManagerTestBuilder.RangeSectionMapTestBuilder(arch);
         builder.MarkCreated();
         var target = new RSLATestTarget(arch, builder.GetReadContext());
 
@@ -177,7 +45,7 @@ public class RangeSectionMapTests
     [ClassData(typeof(MockTarget.StdArch))]
     public void TestLookupOne(MockTarget.Architecture arch)
     {
-        var builder = new Builder(arch);
+        var builder = new ExecutionManagerTestBuilder.RangeSectionMapTestBuilder(arch);
         var inputPC = new TargetCodePointer(0x007f_0000);
         var length = 0x1000u;
         var value = 0x0a0a_0a0au;
